Added Keep Alive Response Checking
I wanted to add the ability to actually check the response of the keep alive polling.

Like other jQuery-idle-Timeout versions (specifically @ehynds 's version https://github.com/ehynds/jquery-idle-timeout)

Added variable 'keepAliveResponse'
If 'keepAliveResponse' is set to false, response is ignored/disregarded (like it already was prior to this PR).
If 'keepAliveResponse' is set to a string (eg. 'OK') then the ajax response from keepAliveUrl is verified against the string. If its different, then the user is redirected to 'keepAliveBadUrl'

Added variable 'keepAliveBadUrl' to set a given URL to redirect to if the above 'keepAliveResponse' is set to anything other than false.

This is useful if you are using the polling of keepAliveUrl to check on settings (in my example, I verify the user's session still exists, if it does I return "OK" if it doesn't I return "ERROR" prompting the redirect to the login page). You could also use this to check on other critical information and redirect to an error page. keepAliveBadUrl could even be modified to show a model dialog window or something.